requirements.txt
apache-airflow
yfinance
pandas
numpy
